Merge ImgLib2 examples with Game of Death example
This migrates all the ImgLib2 examples into /imglib2/examples, as a
single easy-to-use project.

To be consistent with the wiki page, we stick with packageless classes
(except for the Game of Death, which remains in the "game" package).
This results in fewer nested directories, which may be easier for some
new developers. It also discourages reuse of the examples as-is, instead
encouraging adaptation to one's own needs (which is good in this case).