Examples: update ImgOpener construction
Changed the ImgOpener construction to avoid using the non-basic
openImg signature in most cases.

There was no need to always pass ArrayImgFactory, as typically
ImgOptions should be used to specify a factory type (if a common type)
to allow fallback to Cell/Planar if necessary. The ImgFactory signature
is primarily to allow custom factory use. This is now explained in the
example documentation.

Similarly, FloatType was removed from openImg signatures where possible.
In most cases it remains, however, as it is helpful to perform image
processing algorithms on FloatTypes.

Currently these changes do result in a number of new rawtype warnings.
However, there is a ticket on the SCIFIO trac to resolve these.